Make the GitPuller an abstract class extended by GithubPuller
